+++ usr.sbin/bhyve/mmio/devemu_virtio.c
@@ -457,7 +457,7 @@ vq_endchains(struct vqueue_info *vq, int used_all_avail)
  */
 uint64_t
 vi_devemu_read(struct vmctx *ctx, int vcpu, struct devemu_inst *di,
-	       int baridx, uint64_t offset, size_t size)
+	       int baridx, uint64_t offset, int size)
 {
 	struct virtio_softc *vs = di->di_arg;
 	struct virtio_consts *vc;
@@ -554,7 +554,7 @@ done:
  */
 void
 vi_devemu_write(struct vmctx *ctx, int vcpu, struct devemu_inst *di,
-		int baridx, uint64_t offset, size_t size, uint64_t value)
+		int baridx, uint64_t offset, int size, uint64_t value)
 {
 	struct virtio_softc *vs = di->di_arg;
 	struct vqueue_info *vq;
